Refactor blobhelper.py into a package.
It was getting huge.
