+++ src/relstorage/blobhelper/util.py
@@ -1,0 +1,32 @@
+##############################################################################
+#
+# Copyright (c) 2009,2019 Zope Foundation and Contributors.
+# All Rights Reserved.
+#
+# This software is subject to the provisions of the Zope Public License,
+# Version 2.1 (ZPL).  A copy of the ZPL should accompany this distribution.
+# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY AND ALL EXPRESS OR IMPLIED
+# WARRANTIES ARE DISCLAIMED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
+# FOR A PARTICULAR PURPOSE.
+#
+##############################################################################
+from __future__ import absolute_import
+from __future__ import print_function
+
+import os
+import time
+
+import zc.lockfile
+
+def lock_blob(path, retries=6000):
+    lockfilename = os.path.join(os.path.dirname(path), '.lock')
+    n = 0
+    while 1:
+        try:
+            return zc.lockfile.LockFile(lockfilename)
+        except zc.lockfile.LockError:
+            n += 1
+            if n > retries:
+                raise
+            time.sleep(0.01)
